Rename `PantsRunIntegrationTest` to `PantsIntegrationTest`
[ci skip-build-wheels]
[ci skip-rust]
