+++ src/python/pants/base/exiter_integration_test.py
@@ -1,10 +1,10 @@
 # Copyright 2018 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testutil.pants_run_integration_test import PantsRunIntegrationTest, ensure_daemon
+from pants.testutil.pants_integration_test import PantsIntegrationTest, ensure_daemon
 
 
-class ExiterIntegrationTest(PantsRunIntegrationTest):
+class ExiterIntegrationTest(PantsIntegrationTest):
     """Tests that "interesting" exceptions are properly rendered."""
 
     @ensure_daemon
@@ -13,4 +13,4 @@ class ExiterIntegrationTest(PantsRunIntegrationTest):
         pants_run = self.run_pants(["run", "testprojects/src/python/unicode/compilation_failure"])
         self.assert_failure(pants_run)
 
-        self.assertIn("import sys¡", pants_run.stderr_data)
+        self.assertIn("import sys¡", pants_run.stderr)
