+++ src/python/pants/help/list_goals_integration_test.py
@@ -1,18 +1,18 @@
 # Copyright 2019 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testutil.pants_run_integration_test import PantsRunIntegrationTest
+from pants.testutil.pants_integration_test import PantsIntegrationTest
 
 
-class TestListGoalsIntegration(PantsRunIntegrationTest):
+class TestListGoalsIntegration(PantsIntegrationTest):
     def test_goals(self) -> None:
         command = ["goals"]
         pants_run = self.run_pants(command=command)
         self.assert_success(pants_run)
-        assert "to get help for a particular goal" in pants_run.stdout_data
+        assert "to get help for a particular goal" in pants_run.stdout
         # Spot check a few core goals.
         for goal in ["filedeps", "list", "roots", "validate"]:
-            assert goal in pants_run.stdout_data
+            assert goal in pants_run.stdout
 
     def test_only_show_implemented_goals(self) -> None:
         # Some core goals, such as `./pants test`, require downstream implementations to work
@@ -23,7 +23,7 @@ class TestListGoalsIntegration(PantsRunIntegrationTest):
         not_implemented_run = self.run_pants(["--backend-packages=[]", *command,])
         self.assert_success(not_implemented_run)
         for goal in goals_that_need_implementation:
-            assert goal not in not_implemented_run.stdout_data
+            assert goal not in not_implemented_run.stdout
 
         implemented_run = self.run_pants(
             [
@@ -33,11 +33,11 @@ class TestListGoalsIntegration(PantsRunIntegrationTest):
         )
         self.assert_success(implemented_run)
         for goal in goals_that_need_implementation:
-            assert goal in implemented_run.stdout_data
+            assert goal in implemented_run.stdout
 
     def test_ignored_args(self) -> None:
         # Test that arguments (some of which used to be relevant) are ignored.
         command = ["goals", "--all", "--graphviz", "--llama"]
         pants_run = self.run_pants(command=command)
         self.assert_success(pants_run)
-        assert "to get help for a particular goal" in pants_run.stdout_data
+        assert "to get help for a particular goal" in pants_run.stdout
